Add EMG to tracking results
